+++ example/items_example.dart
@@ -1,56 +1,76 @@
 import 'package:jsonize/jsonize.dart';
 
-class Item implements Jsonizable<Item> {
+class Item extends Clonable<Item> {
+  String get cCode => "code";
+
   final String code;
 
   Item(this.code);
   factory Item.empty() => Item("");
+  static CloneFields getFields(Item o) => CloneFields(
+      [CloneField(o.cCode, getter: () => o.code, setter: (v) => {})]);
 
-  // Jsonizable implementation
+  @override
+  String toString() => code;
+
+  // Clonable implementation
   @override
   String get jsonClassCode => "item";
+
   @override
-  Map<String, dynamic> toJson() => {"code": code};
+  Item create(json) => Item(json[cCode]);
+
   @override
-  Item? fromJson(value) => Item(value["code"]);
+  CloneFields get fields => Item.getFields(this);
 }
 
 class DateTimeItem extends Item {
   DateTime dt;
-  DateTimeItem(String code, this.dt) : super(code);
-  factory DateTimeItem.empty() => DateTimeItem("", DateTime(0));
+  DateTimeItem(String code, [DateTime? dt])
+      : dt = dt ?? DateTime(0),
+        super(code);
+  factory DateTimeItem.empty() => DateTimeItem("");
 
-  // Jsonizable implementation
+  @override
+  String toString() => "$code - $dt";
+
+  // Clonable implementation
   @override
   String get jsonClassCode => "dtItem";
 
   @override
-  Map<String, dynamic> toJson() => super.toJson()..["dt"] = dt;
+  DateTimeItem create(json) => DateTimeItem(json[cCode]);
 
   @override
-  DateTimeItem? fromJson(value) => DateTimeItem(value["code"], value["dt"]);
+  CloneFields get fields => Item.getFields(this)
+    ..add(CloneField<DateTime>("dt", getter: () => dt, setter: (v) => dt = v));
 }
 
 class ColorItem extends Item {
-  final int r;
-  final int g;
-  final int b;
-  ColorItem(String code, this.r, this.g, this.b) : super(code);
-  factory ColorItem.empty() => ColorItem("", 0, 0, 0);
+  int r, g, b;
+  ColorItem(String code, {this.r = 0, this.g = 0, this.b = 0}) : super(code);
+  factory ColorItem.empty() => ColorItem("");
 
-  // Jsonizable implementation
+  @override
+  String toString() => "$code - $r.$g.$b";
+
+  // Clonable implementation
   @override
   String get jsonClassCode => "colItem";
 
   @override
-  Map<String, dynamic> toJson() => super.toJson()
-    ..["r"] = r
-    ..["g"] = g
-    ..["b"] = b;
+  ColorItem create(json) => ColorItem(json[cCode]);
 
   @override
-  ColorItem? fromJson(value) =>
-      ColorItem(value["code"], value["r"], value["g"], value["b"]);
+  CloneFields get fields => Item.getFields(this)
+    ..addAll([
+      CloneField<int>("r",
+          getter: () => r, setter: (v) => r = v, defaultValue: 0),
+      CloneField<int>("g",
+          getter: () => g, setter: (v) => g = v, defaultValue: 0),
+      CloneField<int>("b",
+          getter: () => b, setter: (v) => b = v, defaultValue: 0)
+    ]);
 }
 
 void main() {
@@ -58,15 +78,18 @@ void main() {
   Jsonize.registerClasses(
       [Item.empty(), DateTimeItem.empty(), ColorItem.empty()]);
 
-  Map<String, dynamic> myMap = {
-    "item": Item("A base item"),
-    "dt_item": DateTimeItem("Now", DateTime.now()),
-    "color_item": ColorItem("Red", 255, 0, 0)
-  };
+  List myList = [
+    Item("A base item"),
+    DateTimeItem("Now", DateTime.now()),
+    DateTimeItem("A Date", DateTime(2022, 4, 20)),
+    ColorItem("Red", r: 255),
+    ColorItem("Blue", b: 255),
+    ColorItem("Gray", r: 128, g: 128, b: 128)
+  ];
 
-  var jsonRep = Jsonize.toJson(myMap,
+  var jsonRep = Jsonize.toJson(myList,
       jsonClassToken: "!", dateTimeFormat: DateTimeFormat.epoch);
-  var hereIsMyMap = Jsonize.fromJson(jsonRep,
+  var backToLife = Jsonize.fromJson(jsonRep,
       jsonClassToken: "!", dateTimeFormat: DateTimeFormat.epoch);
-  print(hereIsMyMap);
+  print(backToLife);
 }