+++ lib/src/helpers.dart
@@ -1,4 +1,4 @@
-import 'jsonize.dart';
+import '../jsonize.dart';
 
 /// The class used to register [DateTime] type serialization
 class DateTimeJsonable {
@@ -22,7 +22,8 @@ class DateTimeJsonable {
       case DateTimeFormat.epochWithMicros:
         return v.microsecondsSinceEpoch;
       default:
-        throw JsonizeException("DateTime->toJson: '$format' DateTime format!");
+        throw JsonizeException(
+            "DateTimeJsonable", "toJson: '$format' DateTime format!");
     }
   }
 
@@ -34,25 +35,29 @@ class DateTimeJsonable {
         if (v is String) {
           return DateTime.tryParse(v.toString());
         }
-        throw JsonizeException("DateTime->fromJson: String value expected!");
+        throw JsonizeException(
+            "DateTimeJsonable", "fromJson: String value expected!");
       case DateTimeFormat.epoch:
         if (v is int) {
           return DateTime.fromMillisecondsSinceEpoch(v * 1000, isUtc: false);
         }
-        throw JsonizeException("DateTime->fromJson: Integer value expected!");
+        throw JsonizeException(
+            "DateTimeJsonable", "fromJson: Integer value expected!");
       case DateTimeFormat.epochWithMillis:
         if (v is int) {
           return DateTime.fromMillisecondsSinceEpoch(v, isUtc: false);
         }
-        throw JsonizeException("DateTime->fromJson: Integer value expected!");
+        throw JsonizeException(
+            "DateTimeJsonable", "fromJson: Integer value expected!");
       case DateTimeFormat.epochWithMicros:
         if (v is int) {
           return DateTime.fromMicrosecondsSinceEpoch(v, isUtc: false);
         }
-        throw JsonizeException("DateTime->fromJson: Integer value expected!");
+        throw JsonizeException(
+            "DateTimeJsonable", "fromJson: Integer value expected!");
       default:
         throw JsonizeException(
-            "DateTime->fromJson: '$format' DateTime format!)");
+            "DateTimeJsonable", "fromJson: '$format' DateTime format!)");
     }
   }
 
@@ -60,8 +65,8 @@ class DateTimeJsonable {
     if (num >= 0) {
       return num.toString().padLeft(pad, "0");
     }
-    throw JsonizeException(
-        "DateTime->padDigits: DateTime does not handle nevative values!");
+    throw JsonizeException("DateTimeJsonable",
+        "padDigits: DateTime does not handle nevative values!");
   }
 
   String _toString(DateTime v, bool withMillis) {
@@ -84,7 +89,9 @@ class ConvertInfo {
   final Type classType;
   final String jsonClassCode;
   final dynamic convert;
-  ConvertInfo(this.classType, this.jsonClassCode, this.convert);
+  final Jsonizable? emptyObj; // Null if registering a type
+  ConvertInfo(this.classType, this.jsonClassCode, this.convert,
+      [this.emptyObj]);
 }
 
 /// The session class used to store conversion parameters like jsonClassToken.
@@ -133,7 +140,8 @@ class JsonizeSession {
         var convertInfo = _decoders[classType];
         if (convertInfo != null) {
           if (convertCallback != null) {
-            value = convertCallback!(convertInfo.classType, value);
+            value = convertCallback!(
+                convertInfo.classType, value, convertInfo.emptyObj);
           }
           return convertInfo.convert(value);
         }
@@ -154,7 +162,8 @@ class JsonizeSession {
         object is Jsonizable ? object.toJson() : info.convert(object);
     return convertCallback == null
         ? jsonObj
-        : convertCallback!(info.classType, jsonObj);
+        : convertCallback!(
+            info.classType, jsonObj, object is Jsonizable ? object : null);
   }
 
   /// Makes a class token string